Update kinova site_config xacro and base srdf
Add updated arguments to kinova site_config xacros so the model can be
loaded using the latest xacro macros from the ros2_kortex repo.

Always add the manual_grasp_link with the end_effector_link as the
parent. Before, it was added with the tool_frame link as its parent link
which broke the description sometimes since the tool_frame link was only
included if the gripper argument was empty.

Add in disable_collision tags between the robotiq links and the wrist
camera links in the base_config srdf. Without these additions, planning
always fails.