Update kinova base_config control config
Add the streaming_controller back in and change it to joint-based
control rather than twist-based control. Without the streaming
controller defined in the control config an Error is printed to the log
on studio startup since it is referenced in the base_config.yaml file

Update the fault_controller type.

Update the robotiq_gripper_controller joint parameter to the correct
joint.