Set initial joint positions during render
Pass the default values in the `config/initial_positions.yaml` file to
the initial_positions parameter of the load_robot macro during render of
the kinova to set it to the home position by default.

Point to updated ros2_kortex submodule commit that enables the
initial_positions parameter in the load_robot macro.

Point to latest ros2_robotiq_gripper submodule commit.